refactor: 관심상권 상권 분석페이지에서 잘 되도록 변경

+++ total-repo/front/src/views/InterestAreas.vue
@@ -30,6 +30,9 @@
         </div>
       </div>
     </div>
+
+    <!-- Login Modal -->
+    <LoginModal v-if="showLoginPopup" />
   </div>
 </template>
 
@@ -37,10 +40,19 @@
 import { ref, onMounted, onUnmounted } from 'vue';
 import { useRouter } from 'vue-router';
 import { api } from "@/lib/api.js"
+import LoginModal from "@/components/login/LoginModal.vue";
+import { useAccountStore } from "@/stores/useAccountStore";
 
+const accountStore = useAccountStore(); // Use the account store
 const favoriteAreas = ref([]);
+const showLoginPopup = ref(false); // Flag for showing login modal
 
 const fetchFavoriteAreas = async () => { 
+  if (!accountStore.isAuthenticated) {
+    showLoginPopup.value = true;
+    return;
+  }
+
   try {
     const response = await api.get(`/api/favorite/areas/list`);
     if (response.data.area_list) {
@@ -58,6 +70,11 @@ const fetchFavoriteAreas = async () => {
 };
 
 const toggleFavorite = async (area) => {
+  if (!accountStore.isAuthenticated) {
+    showLoginPopup.value = true;
+    return;
+  }
+
   try {
     if (area.isFavorite) {
       await api.delete(`/api/favorite/areas/${area.favorite_id}`);
@@ -78,15 +95,21 @@ const navigateToMarketAnalysis = (area) => {
 };
 
 const handleModalShown = () => {
-  fetchFavoriteAreas();
+  if (!accountStore.isAuthenticated) {
+    showLoginPopup.value = true;
+  }
+  else{
+    fetchFavoriteAreas();
+  }
 };
 
+//컴포넌트가 DOM에 마운트된 직후(즉, 화면에 표시된 직후)에 실행됩니다.
 onMounted(() => {
-  // Add event listener for modal shown event
   const modalElement = document.getElementById('favoriteArea');
   modalElement.addEventListener('shown.bs.modal', handleModalShown);
 });
 
+// 컴포넌트가 DOM에서 제거되기 직전에 실행됩니다.
 onUnmounted(() => {
   // Remove event listener when component is unmounted
   const modalElement = document.getElementById('favoriteArea');
